Extract donutChart fixtures from testData
