+++ src/charts/donut/donutChart.fixtures.js
@@ -1,0 +1,30 @@
+const with4Slices = () => [
+    {
+        quantity: 60,
+        percentage: 60,
+        name: 'React',
+        id: 1,
+    },
+    {
+        quantity: 20,
+        percentage: 20,
+        name: 'Ember',
+        id: 2,
+    },
+    {
+        quantity: 10,
+        percentage: 10,
+        name: 'Angular',
+        id: 3,
+    },
+    {
+        quantity: 10,
+        percentage: 10,
+        name: 'Backbone',
+        id: 4,
+    },
+];
+
+export default {
+    with4Slices,
+};
